# Feature/optimization get avatar state