액션에서 사용하는 GetAvatarState에서 필요한 상태만 가져오도록 변경하여 성능을 개선합니다.

## 성능비교

### SweepAdventureBoss

전 Execute → 1088ms
후 Execute → 1000ms (avatar차이 40)

### Wanted

전 Execute → 210ms
후 Execute → 110ms (avatar차이 100)